services/auth/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.6.0
pymongo==4.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
